Ports the strongSwan core libraries and some plugins to the Windows platform
using a MinGW based toolchain. Beside generic platform abstraction and
the windows.[ch] compatibility layer, this merge introduces a Windows native
threading backend and a charon-svc Windows IKE service.

Travis adds a MinGW cross-compile build to Windows, and further enables -Werror
to let builds fail for all compiler warnings with gcc and Clang.

+++ src/libimcv/imv/imv_agent.c
@@ -612,7 +612,7 @@ METHOD(imv_agent_t, reserve_additional_ids, TNC_Result,
 		count--;
 
 		/* store the scalar value in the pointer */
-		pointer = (void*)id;
+		pointer = (void*)(uintptr_t)id;
 		this->additional_ids->insert_last(this->additional_ids, pointer);
 		DBG2(DBG_IMV, "IMV %u \"%s\" reserved additional ID %u",
 					  this->id, this->name, id);
@@ -831,5 +831,3 @@ imv_agent_t *imv_agent_create(const char *name,
 
 	return &this->public;
 }
-
-
